backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymupdf==1.23.8
reportlab==4.1.0
pillow==10.3.0
werkzeug==3.0.1
python-docx==1.1.0
pdf2docx==0.5.8
docx2pdf==0.8.0
